Extract PowerShellProxyFixture from PowerShellFixture
