+++ source/Calamari.Tests/Fixtures/PowerShell/PowerShellFixture.cs
@@ -8,21 +8,17 @@ using Calamari.Integration.FileSystem;
 using Calamari.Integration.Scripting;
 using Calamari.Tests.Helpers;
 using Calamari.Util.Environments;
+using FluentAssertions;
 using Newtonsoft.Json;
 using NUnit.Framework;
 using Octostache;
+using SetProxy;
 
 namespace Calamari.Tests.Fixtures.PowerShell
 {
     [TestFixture]
     public class PowerShellFixture : CalamariFixture
     {
-        [TearDown]
-        public void Teardown()
-        {
-            ResetProxyEnvironmentVariables();
-        }
-
         [Test]
         [Platform]
         [Category(TestCategory.CompatibleOS.Windows)]
@@ -357,7 +353,6 @@ namespace Calamari.Tests.Fixtures.PowerShell
             output.AssertOutput("V3= GHI");
             output.AssertOutput("FooBar= Hello"); // Legacy - '_' used to be removed
             output.AssertOutput("Foo_Bar= Hello"); // Current - '_' is valid in PowerShell
-
         }
 
         [Test]
@@ -572,148 +567,6 @@ namespace Calamari.Tests.Fixtures.PowerShell
                 output.AssertSuccess();
                 output.AssertOutput(CalamariEnvironment.IsRunningOnWindows ? "Hello Powershell" : "Hello Bash");
             }
-        }
-
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxyNotSet_ShouldNotSetEnvironmentVariables()
-        {
-            ResetProxyEnvironmentVariables();
-
-            EnvironmentHelper.SetEnvironmentVariable("TentacleUseDefaultProxy", false.ToString());
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyHost", "");
-
-            var (output, _) = RunScript("Proxy.ps1");
-
-            output.AssertSuccess();
-            output.AssertNoOutput($"Setting Proxy Environment Variables");
-            output.AssertOutputContains($"HTTP_PROXY: ");
-            output.AssertOutputContains($"HTTPS_PROXY: ");
-        }
-
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxyConfigured_ShouldSetEnvironmentVariables()
-        {
-            ResetProxyEnvironmentVariables();
-
-            var proxyHost = "hostname";
-            var proxyPort = "3456";
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyHost", proxyHost);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPort", proxyPort);
-
-            var (output, _) = RunScript("Proxy.ps1");
-
-            output.AssertSuccess();
-            output.AssertOutputContains($"HTTP_PROXY: http://{proxyHost}:{proxyPort}");
-            output.AssertOutputContains($"HTTPS_PROXY: http://{proxyHost}:{proxyPort}");
-        }
-
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxyWithAuthConfigured_ShouldSetEnvironmentVariables()
-        {
-            ResetProxyEnvironmentVariables();
-
-            var proxyHost = "hostname";
-            var proxyPort = "3456";
-            var proxyUsername = "username";
-            var proxyPassword = "password";
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyHost", proxyHost);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPort", proxyPort);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyUsername", proxyUsername);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPassword", proxyPassword);
-
-            var (output, _) = RunScript("Proxy.ps1");
-
-            output.AssertSuccess();
-            output.AssertOutputContains($"HTTP_PROXY: http://{proxyUsername}:{proxyPassword}@{proxyHost}:{proxyPort}");
-            output.AssertOutputContains($"HTTPS_PROXY: http://{proxyUsername}:{proxyPassword}@{proxyHost}:{proxyPort}");
-        }
-
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxyEnvironmentAlreadyConfigured_ShouldSetNotSetVariables()
-        {
-            ResetProxyEnvironmentVariables();
-
-            var proxyHost = "hostname";
-            var proxyPort = "3456";
-            var httpProxy = "http://proxy:port";
-            var httpsProxy = "http://proxy2:port";
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyHost", proxyHost);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPort", proxyPort);
-            EnvironmentHelper.SetEnvironmentVariable("HTTP_PROXY", httpProxy);
-            EnvironmentHelper.SetEnvironmentVariable("HTTPS_PROXY", httpsProxy);
-
-            var (output, _) = RunScript("Proxy.ps1");
-
-            output.AssertSuccess();
-            output.AssertOutputContains($"HTTP_PROXY: {httpProxy}");
-            output.AssertOutputContains($"HTTPS_PROXY: {httpsProxy}");
-
-            ResetProxyEnvironmentVariables();
-        }
-
-#if NETFX
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxySetToSystem_ShouldSetVariablesCorrectly()
-        {
-            ResetProxyEnvironmentVariables();
-
-            var (output, _) = RunScript("Proxy.ps1");
-            var systemProxyUri = System.Net.WebRequest.GetSystemWebProxy().GetProxy(new Uri(@"https://octopus.com"));
-            if (systemProxyUri.Host == "octopus.com")
-            {
-                output.AssertSuccess();
-                output.AssertOutputContains($"HTTP_PROXY: ");
-                output.AssertOutputContains($"HTTPS_PROXY: ");
-            }
-            else
-            {
-                output.AssertSuccess();
-                output.AssertOutputContains($"HTTP_PROXY: http://{systemProxyUri.Host}:{systemProxyUri.Port}");
-                output.AssertOutputContains($"HTTPS_PROXY: http://{systemProxyUri.Host}:{systemProxyUri.Port}");
-            }
-        }
-#endif
-
-#if NETFX
-        [Test]
-        [Category(TestCategory.CompatibleOS.Windows)]
-        public void ProxyNoConfig_ShouldSetNotSetVariables()
-        {
-            ResetProxyEnvironmentVariables();
-
-            var (output, _) = RunScript("Proxy.ps1");
-            var systemProxyUri = System.Net.WebRequest.GetSystemWebProxy().GetProxy(new Uri(@"http://octopus.com"));
-            if (systemProxyUri.Host == "octopus.com")
-            {
-                output.AssertSuccess();
-                output.AssertOutputContains($"HTTP_PROXY: ");
-                output.AssertOutputContains($"HTTPS_PROXY: ");
-            }
-            else
-            {
-                output.AssertSuccess();
-                output.AssertOutputContains($"HTTP_PROXY: http://{systemProxyUri.Host}:{systemProxyUri.Port}");
-                output.AssertOutputContains($"HTTPS_PROXY: http://{systemProxyUri.Host}:{systemProxyUri.Port}");
-            }
-        }
-#endif
-
-
-        private void ResetProxyEnvironmentVariables()
-        {
-            EnvironmentHelper.SetEnvironmentVariable("TentacleUseDefaultProxy", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyHost", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPort", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyUsername", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("TentacleProxyPassword", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("HTTP_PROXY", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("HTTPS_PROXY", string.Empty);
-            EnvironmentHelper.SetEnvironmentVariable("NO_PROXY", string.Empty);
         }
     }
 }
